+++ packages/paste-website/src/components/site-wrapper/sidebar/SidebarItem.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
-import {Box, BoxProps} from '@twilio-paste/box';
+import {Box} from '@twilio-paste/box';
 
-export const SidebarItem: React.FC<BoxProps> = (props) => {
+interface SidebarItemProps {
+  children: React.ReactNode;
+}
+
+export const SidebarItem: React.FC<SidebarItemProps> = ({children}) => {
   return (
-    <Box
-      as="li"
-      margin="space20"
-      fontSize={['fontSize50', 'fontSize50', 'fontSize30']}
-      lineHeight="lineHeight20"
-      {...props}
-    />
+    <Box as="li" fontSize={['fontSize50', 'fontSize50', 'fontSize30']} lineHeight="lineHeight20">
+      {children}
+    </Box>
   );
 };