+++ packages/paste-website/src/utils/GraphqlUtils.ts
@@ -1,5 +1,5 @@
 import * as _ from 'lodash';
-import {getHumanizedNameFromPackageName} from './RouteUtils';
+import {getHumanizedNameFromPackageName, getNameFromPackageName} from './RouteUtils';
 import {sentenceCase} from './SentenceCase';
 
 export type GraphqlData = Record<string, any>;
@@ -23,13 +23,13 @@ const getPath = (key: string): string => {
     case 'allPasteDesignTokensPackage':
     case 'allPasteLayout':
     case 'allPasteLibraries':
-    case 'allPastePattern':
     case 'allPastePrimitive':
     case 'allPasteThemePackage': {
       return 'edges[0].node';
     }
 
-    case 'allAirtable': {
+    case 'allAirtable':
+    case 'allPastePattern': {
       return 'edges[0].node.data';
     }
 
@@ -103,4 +103,54 @@ export const getNormalizedHeaderData = (data: GraphqlData): GraphqlData => {
       [currentKey]: mutatedData,
     };
   }, {});
+};
+
+const getNavigationDataMutation = (key: string): MutationFunction => {
+  switch (key) {
+    case 'allPasteComponent':
+    case 'allPasteLayout':
+    case 'allPastePrimitive': {
+      return ({name, status}: GraphqlData): GraphqlData => ({
+        name: getHumanizedNameFromPackageName(name),
+        packageName: name,
+        slug: getNameFromPackageName(name),
+        packageStatus: status,
+      });
+    }
+
+    case 'allPastePattern': {
+      return ({data}: GraphqlData): GraphqlData => {
+        const {status, Feature} = data;
+        return {
+          name: Feature,
+          packageStatus: status,
+          slug: Feature.toLowerCase().replace(/\s/g, '-'),
+        };
+      };
+    }
+
+    default:
+      return (data: GraphqlData) => data;
+  }
+};
+
+export const getNormalizedNavigationData = (data: GraphqlData): GraphqlData => {
+  const queryKeys = Object.keys(data);
+  const normalizedData: Record<string, GraphqlData> = {};
+
+  queryKeys.forEach((currentKey) => {
+    if (shouldFlatten(currentKey)) {
+      const dataFragment = data[currentKey].edges;
+      const mutateOperation = getNavigationDataMutation(currentKey);
+      let res: Array<GraphqlData> = [];
+
+      dataFragment.forEach((currentPackage: GraphqlData) => {
+        const mutatedData = mutateOperation(currentPackage.node);
+        res = [...res, mutatedData];
+      });
+      normalizedData[currentKey] = res;
+    }
+  });
+
+  return normalizedData;
 };